backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
pytest>=8.0
pytest-django>=4.8
